+++ TombersAndPyramiders/Game/GameObjects/MiscSquare.cpp
