+++ TombersAndPyramiders/Game/GameObjects/MiscSquare.h
