--- TombersAndPyramiders/GarbageCollection.cpp
@@ -1,4 +1,0 @@
-#include "GarbageCollection.h"
-
-bool isBeingCollected = false;
-bool doneBeingCollected = true;
