--- TombersAndPyramiders/GarbageCollection.h
@@ -1,4 +1,0 @@
-#pragma once
-
-extern bool isBeingCollected;
-extern bool doneBeingCollected;
