--- TombersAndPyramiders/Sound.cpp
@@ -1,10 +1,0 @@
-#include "Sound.h"
-
-
-Sound::Sound() : Component(gameObject)
-{
-}
-
-Sound::~Sound()
-{
-}
