--- TombersAndPyramiders/Sound.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "Component.h"
-
-
-class Sound : Component {
-private:
-	Sound::Sound();
-
-	Sound::~Sound();
-public:
-};
